# Make Olympus decompressor a bit faster (-12%)