--- a/src/librawspeed/decompressors/OlympusDecompressor.cpp
+++ b/src/librawspeed/decompressors/OlympusDecompressor.cpp
@@ -22,32 +22,53 @@
 
 #include "decompressors/OlympusDecompressor.h"
 #include "adt/Array2DRef.h"
+#include "adt/Bit.h"
 #include "adt/Casts.h"
 #include "adt/Invariant.h"
 #include "adt/Point.h"
 #include "bitstreams/BitStreamerMSB.h"
 #include "common/RawImage.h"
+#include "common/SimpleLUT.h"
 #include "decoders/RawDecoderException.h"
+#include "decompressors/AbstractDecompressor.h"
 #include "io/ByteStream.h"
+#include <algorithm>
 #include <array>
 #include <cmath>
+#include <cstddef>
 #include <cstdint>
 #include <cstdlib>
 #include <utility>
 
 namespace rawspeed {
 
-OlympusDecompressor::OlympusDecompressor(RawImage img) : mRaw(std::move(img)) {
-  if (mRaw->getCpp() != 1 || mRaw->getDataType() != RawImageType::UINT16 ||
-      mRaw->getBpp() != sizeof(uint16_t))
-    ThrowRDE("Unexpected component count / data type");
+namespace {
 
-  const uint32_t w = mRaw->dim.x;
-  const uint32_t h = mRaw->dim.y;
+class OlympusDecompressorImpl final : public AbstractDecompressor {
+  RawImage mRaw;
 
-  if (w == 0 || h == 0 || w % 2 != 0 || h % 2 != 0 || w > 10400 || h > 7792)
-    ThrowRDE("Unexpected image dimensions found: (%u; %u)", w, h);
-}
+  // A table to quickly look up "high" value
+  const SimpleLUT<int8_t, 12> bittable{
+      [](size_t i, [[maybe_unused]] unsigned tableSize) {
+        int high;
+        for (high = 0; high < 12; high++)
+          if (extractHighBits(i, high, /*effectiveBitwidth=*/11) & 1)
+            break;
+        return std::min(12, high);
+      }};
+
+  inline __attribute__((always_inline)) int
+  parseCarry(BitStreamerMSB& bits, std::array<int, 3>& carry) const;
+
+  static inline int getPred(Array2DRef<uint16_t> out, int row, int col);
+
+  void decompressRow(BitStreamerMSB& bits, int row) const;
+
+public:
+  explicit OlympusDecompressorImpl(RawImage img) : mRaw(std::move(img)) {}
+
+  void decompress(ByteStream input) const;
+};
 
 /* This is probably the slowest decoder of them all.
  * I cannot see any way to effectively speed up the prediction
@@ -57,8 +78,8 @@ OlympusDecompressor::OlympusDecompressor(RawImage img) : mRaw(std::move(img)) {
  */
 
 inline __attribute__((always_inline)) int
-OlympusDecompressor::parseCarry(BitStreamerMSB& bits,
-                                std::array<int, 3>& carry) const {
+OlympusDecompressorImpl::parseCarry(BitStreamerMSB& bits,
+                                    std::array<int, 3>& carry) const {
   bits.fill();
   int i = 2 * (carry[2] < 3);
   int nbits;
@@ -85,8 +106,8 @@ OlympusDecompressor::parseCarry(BitStreamerMSB& bits,
   return (diff * 4) | low;
 }
 
-inline int OlympusDecompressor::getPred(const Array2DRef<uint16_t> out, int row,
-                                        int col) {
+inline int OlympusDecompressorImpl::getPred(const Array2DRef<uint16_t> out,
+                                            int row, int col) {
   auto getLeft = [&]() { return out(row, col - 2); };
   auto getUp = [&]() { return out(row - 2, col); };
   auto getLeftUp = [&]() { return out(row - 2, col - 2); };
@@ -120,7 +141,8 @@ inline int OlympusDecompressor::getPred(const Array2DRef<uint16_t> out, int row,
   return pred;
 }
 
-void OlympusDecompressor::decompressRow(BitStreamerMSB& bits, int row) const {
+void OlympusDecompressorImpl::decompressRow(BitStreamerMSB& bits,
+                                            int row) const {
   invariant(mRaw->dim.y > 0);
   invariant(mRaw->dim.x > 0);
   invariant(mRaw->dim.x % 2 == 0);
@@ -141,7 +163,7 @@ void OlympusDecompressor::decompressRow(BitStreamerMSB& bits, int row) const {
   }
 }
 
-void OlympusDecompressor::decompress(ByteStream input) const {
+void OlympusDecompressorImpl::decompress(ByteStream input) const {
   invariant(mRaw->dim.y > 0);
   invariant(mRaw->dim.x > 0);
   invariant(mRaw->dim.x % 2 == 0);
@@ -151,6 +173,24 @@ void OlympusDecompressor::decompress(ByteStream input) const {
 
   for (int y = 0; y < mRaw->dim.y; y++)
     decompressRow(bits, y);
+}
+
+} // namespace
+
+OlympusDecompressor::OlympusDecompressor(RawImage img) : mRaw(std::move(img)) {
+  if (mRaw->getCpp() != 1 || mRaw->getDataType() != RawImageType::UINT16 ||
+      mRaw->getBpp() != sizeof(uint16_t))
+    ThrowRDE("Unexpected component count / data type");
+
+  const uint32_t w = mRaw->dim.x;
+  const uint32_t h = mRaw->dim.y;
+
+  if (w == 0 || h == 0 || w % 2 != 0 || h % 2 != 0 || w > 10400 || h > 7792)
+    ThrowRDE("Unexpected image dimensions found: (%u; %u)", w, h);
+}
+
+void OlympusDecompressor::decompress(ByteStream input) const {
+  OlympusDecompressorImpl(mRaw).decompress(input);
 }
 
 } // namespace rawspeed