Since it's still sees real-world use
in newly released cameras
(OM SYSTEM OM1 Mark II, e.g.),
i guess it's performance matters.

As comment notes, that predictor thingy
is really horrifying and slow,
but we can improve things somewhat...

```
raw.pixls.us-unique/Olympus/XZ-1$ /repositories/googlebenchmark/tools/compare.py -a benchmarks ~/rawspeed/build-Clang17-release/src/utilities/rsbench/rsbench{-old,} p1319978.orf --benchmark_repetitions=49
RUNNING: /home/lebedevri/rawspeed/build-Clang17-release/src/utilities/rsbench/rsbench-old p1319978.orf --benchmark_repetitions=49 --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmpv0l1_ew2
2024-02-29T21:01:02+03:00
Running /home/lebedevri/rawspeed/build-Clang17-release/src/utilities/rsbench/rsbench-old
Run on (32 X 3399.15 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 512 KiB (x16)
  L3 Unified 32768 KiB (x2)
Load Average: 0.51, 0.71, 0.80
----------------------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------
p1319978.orf/threads:32/process_time/real_time_mean          108 ms          108 ms           49 CPUTime,s=0.107614 CPUTime/WallTime=0.999943 Pixels=10.1568M Pixels/CPUTime=94.3818M Pixels/WallTime=94.3764M Raws/CPUTime=9.29247 Raws/WallTime=9.29194 WallTime,s=0.10762
p1319978.orf/threads:32/process_time/real_time_median        108 ms          108 ms           49 CPUTime,s=0.107606 CPUTime/WallTime=0.999966 Pixels=10.1568M Pixels/CPUTime=94.3889M Pixels/WallTime=94.3824M Raws/CPUTime=9.29318 Raws/WallTime=9.29253 WallTime,s=0.107613
p1319978.orf/threads:32/process_time/real_time_stddev      0.077 ms        0.077 ms           49 CPUTime,s=76.5394u CPUTime/WallTime=51.4451u Pixels=0 Pixels/CPUTime=66.9659k Pixels/WallTime=67.538k Raws/CPUTime=6.59321m Raws/WallTime=6.64954m WallTime,s=77.1944u
p1319978.orf/threads:32/process_time/real_time_cv           0.07 %          0.07 %            49 CPUTime,s=0.07% CPUTime/WallTime=0.01% Pixels=0.00% Pixels/CPUTime=0.07% Pixels/WallTime=0.07% Raws/CPUTime=0.07% Raws/WallTime=0.07% WallTime,s=0.07%
RUNNING: /home/lebedevri/rawspeed/build-Clang17-release/src/utilities/rsbench/rsbench p1319978.orf --benchmark_repetitions=49 --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmp85xmyvj3
2024-02-29T21:01:33+03:00
Running /home/lebedevri/rawspeed/build-Clang17-release/src/utilities/rsbench/rsbench
Run on (32 X 3400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 512 KiB (x16)
  L3 Unified 32768 KiB (x2)
Load Average: 0.76, 0.75, 0.81
----------------------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------
p1319978.orf/threads:32/process_time/real_time_mean         94.9 ms         94.9 ms           49 CPUTime,s=0.0949223 CPUTime/WallTime=0.999924 Pixels=10.1568M Pixels/CPUTime=107.001M Pixels/WallTime=106.993M Raws/CPUTime=10.5349 Raws/WallTime=10.5341 WallTime,s=0.0949295
p1319978.orf/threads:32/process_time/real_time_median       94.9 ms         94.9 ms           49 CPUTime,s=0.0949059 CPUTime/WallTime=0.999956 Pixels=10.1568M Pixels/CPUTime=107.02M Pixels/WallTime=107.005M Raws/CPUTime=10.5368 Raws/WallTime=10.5353 WallTime,s=0.0949188
p1319978.orf/threads:32/process_time/real_time_stddev      0.041 ms        0.041 ms           49 CPUTime,s=40.7811u CPUTime/WallTime=63.1188u Pixels=0 Pixels/CPUTime=45.9084k Pixels/WallTime=46.2605k Raws/CPUTime=4.51996m Raws/WallTime=4.55464m WallTime,s=41.1006u
p1319978.orf/threads:32/process_time/real_time_cv           0.04 %          0.04 %            49 CPUTime,s=0.04% CPUTime/WallTime=0.01% Pixels=0.00% Pixels/CPUTime=0.04% Pixels/WallTime=0.04% Raws/CPUTime=0.04% Raws/WallTime=0.04% WallTime,s=0.04%
Comparing /home/lebedevri/rawspeed/build-Clang17-release/src/utilities/rsbench/rsbench-old to /home/lebedevri/rawspeed/build-Clang17-release/src/utilities/rsbench/rsbench
Benchmark                                                               Time             CPU      Time Old      Time New       CPU Old       CPU New
----------------------------------------------------------------------------------------------------------------------------------------------------
p1319978.orf/threads:32/process_time/real_time_pvalue                 0.0000          0.0000      U Test, Repetitions: 49 vs 49
p1319978.orf/threads:32/process_time/real_time_mean                  -0.1179         -0.1179           108            95           108            95
p1319978.orf/threads:32/process_time/real_time_median                -0.1180         -0.1180           108            95           108            95
p1319978.orf/threads:32/process_time/real_time_stddev                -0.4678         -0.4670             0             0             0             0
p1319978.orf/threads:32/process_time/real_time_cv                    -0.3966         -0.3957             0             0             0             0
OVERALL_GEOMEAN                                                      -0.1179         -0.1179             0             0             0             0
```